Move router into zmq.messagebus.router
